Support latest SVD from stm32-rs (except stm32u)
The newer SVD files apparently contain some clusters that dont use the
dimIndex array feature. I have hacked a workaround in stm32ral.py to
make it work anyways. I'm not too familiar with the SVD spec so there is
probably a neater solution than this..

Also I could'nt get it to work for the stm32u series so I left it out
ouf the makefile. When building for stm32u, this is the output I got:

    return list(map(*args))
  File "stm32ral.py", line 1322, in from_svdfile
    return cls.from_svd(svd, device_name)
  File "stm32ral.py", line 1298, in from_svd
    raise e
  File "stm32ral.py", line 1295, in from_svd
    cpu = CPU.from_svd(svd, svd.find('cpu'))
  File "stm32ral.py", line 1133, in from_svd
    name = get_string(node, 'name')
  File "stm32ral.py", line 1657, in get_string
    text = node.findtext(tag, default=default)
AttributeError: 'NoneType' object has no attribute 'findtext'